app/requirements.txt
aniso8601==9.0.1
blinker==1.7.0
click==8.1.7
colorama==0.4.4
coverage==5.5
docker==7.0.0
exceptiongroup==1.2.0
Faker==14.0.0
Flask==2.3.0
Flask-Cors==3.0.10
Flask-JWT==0.2.0
Flask-JWT-Extended==4.6.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
importlib-metadata==7.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.18.0
marshmallow-sqlalchemy==0.25.0
moviepy==1.0.3
packaging==23.2
pika==1.3.2
pluggy==1.3.0
pg8000
PyJWT==2.1.0
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.15
tomli==2.0.1
waitress==3.0.0
Werkzeug==3.0.1
zipp==3.17.0
google-cloud-storage=2.16.0